ejercico4

+++ Ej4/cuenta_banco.cpp
@@ -8,7 +8,7 @@ CuentaBanco::CuentaBanco(string &Titular)
     :titularCuenta(Titular), balance(0) {}
 
 
-void CuentaBanco::Depositar(double& dinero_depositar){
+void CuentaBanco::Depositar(int& dinero_depositar){
     balance += dinero_depositar;
     cout<<"Dinero depositado exitosamente"<<endl;
     return;
